+++ src/utils/defaults.ts
@@ -2,7 +2,7 @@ import { lstatSync, readdirSync, readFileSync, realpathSync } from './fs'; // es
 import { basename, dirname, isAbsolute, resolve } from './path';
 import error from './error';
 import Module from '../Module';
-import relativeId, { jsExts } from './relativeId';
+import relativeId from './relativeId';
 import { InputOptions } from '../rollup/types';
 
 export function load(id: string) {
@@ -29,10 +29,10 @@ function findFile(file: string, preserveSymlinks: boolean): string | void {
 function addJsExtensionIfNecessary(file: string, preserveSymlinks: boolean) {
 	let found = findFile(file, preserveSymlinks);
 	if (found) return found;
-	for (let ext of jsExts) {
-		found = findFile(file + ext, preserveSymlinks);
-		if (found) return found;
-	}
+	found = findFile(file + '.mjs', preserveSymlinks);
+	if (found) return found;
+	found = findFile(file + '.js', preserveSymlinks);
+	return found;
 }
 
 export function resolveId(options: InputOptions) {