+++ test/chunking-form/samples/aliasing-extensions/_expected/cjs/main1.js
@@ -1,0 +1,5 @@
+'use strict';
+
+console.log('main1');
+Promise.resolve(require("./main4.dynamic.js"));
+Promise.resolve(require("./main5.js"));