+++ test/chunking-form/samples/aliasing-extensions/_expected/system/main1.js
@@ -1,0 +1,12 @@
+System.register([], function (exports, module) {
+	'use strict';
+	return {
+		execute: function () {
+
+			console.log('main1');
+			module.import("./main4.dynamic.js");
+			module.import("./main5.js");
+
+		}
+	};
+});