+++ test/form/samples/mjs/_config.js
@@ -1,0 +1,6 @@
+module.exports = {
+	description: 'supports loading mjs with precedence',
+	options: {
+		output: { name: 'myBundle' }
+	}
+};