+++ test/form/samples/mjs/_expected/amd.js
@@ -1,0 +1,12 @@
+define(['exports'], function (exports) { 'use strict';
+
+	var dep = 'js';
+
+	var dep$1 = 'mjs';
+
+	exports.depJs = dep;
+	exports.depMjs = dep$1;
+
+	Object.defineProperty(exports, '__esModule', { value: true });
+
+});