+++ test/form/samples/mjs/_expected/system.js
@@ -1,0 +1,12 @@
+System.register('myBundle', [], function (exports, module) {
+	'use strict';
+	return {
+		execute: function () {
+
+			var dep = exports('depJs', 'js');
+
+			var dep$1 = exports('depMjs', 'mjs');
+
+		}
+	};
+});